[UPDATE] reduce the precision of textures

+++ Final_Indoor/src/GaussianBlurrer.cpp
@@ -17,7 +17,7 @@ void GaussianBlurrer::setupFrameBuffer(glm::ivec2 ws)
     {
         glBindFramebuffer(GL_FRAMEBUFFER, altFbo[i]);
         glBindTexture(GL_TEXTURE_2D, altTex[i]);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, ws.x, ws.y, 0, GL_RGBA, GL_FLOAT, NULL);
+        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, ws.x, ws.y, 0, GL_RGBA, GL_FLOAT, NULL);
 
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
